requirements.txt
asgiref==3.7.2
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1